Added service handler update checker
A service handler can check if it is up-to-date and supports the current service implementation. Example: The firebase service handler checks that the versions listed in the firestore appInfo document matches the versions specified in 'FirebaseServiceHandler.supportedAppInfo'. If any version does not match, the service handler is considered out of date. The host application can decide what to do with this information.